Smart-QR/requirements.txt
Django==4.2.7
qrcode[pil]==7.4.2
Pillow==10.1.0
stripe==7.8.0
python-decouple==3.8